상품 테이블에 할인률 컬럼을 추가한뒤 할인률을 임의로 넣어서 할인이 되는 상품들이 대하여 전부 할인률적용

+++ asset/src/main/java/springBootMVCAsset/command/PurchaseCommand.java
@@ -1,11 +1,13 @@
 package springBootMVCAsset.command;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class PurchaseCommand {
 	String goodsNums;
-	Integer  totalPaymentPrice;
+	BigDecimal  totalPaymentPrice;
 	String purchaseName;
 	String deliveryAddr;
 	String deliveryAddrDetail;
